+++ src/standup.py
@@ -10,7 +10,7 @@ def standup_start(token, channel_id, length):
     then at the end of the standup period a message will be added to the message queue
     in the channel from the user who started the standup.
     Input: token (str), channel_id (int), length (int)
-    Output: time_finish (UNIX timestamp float)
+    Output: time_finish (UNIX timestamp int)
     '''
     # Retrieve data
     auth_user = user_with_token(token)
@@ -43,7 +43,7 @@ def standup_active(token, channel_id):
     and what time the standup finishes. If no standup is active,
     then time_finish returns None.
     Input: token (str), channel_id (int)
-    Output: is_active (bool), time_finish (UNIX timestamp float)
+    Output: is_active (bool), time_finish (UNIX timestamp int)
     '''
     # Retrieve data
     auth_user = user_with_token(token)
